+++ library/include/rocrand/rocrand_mtgp32.h
@@ -1,4 +1,4 @@
-// Copyright (c) 2017-2023 Advanced Micro Devices, Inc. All rights reserved.
+// Copyright (c) 2017-2024 Advanced Micro Devices, Inc. All rights reserved.
 //
 // Permission is hereby granted, free of charge, to any person obtaining a copy
 // of this software and associated documentation files (the "Software"), to deal
@@ -234,25 +234,15 @@ public:
         return this->next();
     }
 
-    FQUALIFIERS
-    unsigned int next()
+    FQUALIFIERS unsigned int next()
     {
-#if defined(__HIP_DEVICE_COMPILE__)
-        unsigned int t   = threadIdx.x;
-        unsigned int d   = blockDim.x;
-        int pos = pos_tbl;
-        unsigned int r;
-        unsigned int o;
-
-        r = para_rec(m_state.status[(t + m_state.offset) & MTGP_MASK],
-                     m_state.status[(t + m_state.offset + 1) & MTGP_MASK],
-                     m_state.status[(t + m_state.offset + pos) & MTGP_MASK]);
-        m_state.status[(t + m_state.offset + MTGP_N) & MTGP_MASK] = r;
-
-        o = temper(r, m_state.status[(t + m_state.offset + pos - 1) & MTGP_MASK]);
+#ifdef __HIP_DEVICE_COMPILE__
+        unsigned int o = next_thread(threadIdx.x);
         __syncthreads();
-        if (t == 0)
-            m_state.offset = (m_state.offset + d) & MTGP_MASK;
+        if(threadIdx.x == 0)
+        {
+            m_state.offset = (m_state.offset + blockDim.x) & MTGP_MASK;
+        }
         __syncthreads();
         return o;
 #else
@@ -323,6 +313,19 @@ private:
         return r;
     }
 
+protected:
+    /// \brief Generate the next value for thread `thread_idx` and modify state in the process,
+    ///   do not update offset.
+    FQUALIFIERS unsigned int next_thread(unsigned int thread_idx)
+    {
+        const unsigned int r
+            = para_rec(m_state.status[(thread_idx + m_state.offset) & MTGP_MASK],
+                       m_state.status[(thread_idx + m_state.offset + 1) & MTGP_MASK],
+                       m_state.status[(thread_idx + m_state.offset + pos_tbl) & MTGP_MASK]);
+        m_state.status[(thread_idx + m_state.offset + MTGP_N) & MTGP_MASK] = r;
+        return temper(r, m_state.status[(thread_idx + m_state.offset + pos_tbl - 1) & MTGP_MASK]);
+    }
+
 public:
     // State
     mtgp32_state m_state;
@@ -355,9 +358,9 @@ typedef rocrand_device::mtgp32_params mtgp32_params;
  * \brief Initializes MTGP32 states
  *
  * Initializes MTGP32 states on the host-side by allocating a state array in host
- * memory, initializes that array, and copies the result to device memory.
+ * memory, initializes that array, and copies the result to device or host memory.
  *
- * \param d_state - Pointer to an array of states in device memory
+ * \param state - Pointer to an array of states in device or host memory
  * \param params - Pointer to an array of type mtgp32_fast_params in host memory
  * \param n - Number of states to initialize
  * \param seed - Seed value
@@ -366,11 +369,10 @@ typedef rocrand_device::mtgp32_params mtgp32_params;
  * - ROCRAND_STATUS_ALLOCATION_FAILED if states could not be initialized
  * - ROCRAND_STATUS_SUCCESS if states are initialized
  */
-__host__ inline
-rocrand_status rocrand_make_state_mtgp32(rocrand_state_mtgp32 * d_state,
-                                         mtgp32_fast_params params[],
-                                         int n,
-                                         unsigned long long seed)
+__host__ inline rocrand_status rocrand_make_state_mtgp32(rocrand_state_mtgp32* state,
+                                                         mtgp32_fast_params    params[],
+                                                         int                   n,
+                                                         unsigned long long    seed)
 {
     int i;
     rocrand_state_mtgp32 * h_state = (rocrand_state_mtgp32 *) malloc(sizeof(rocrand_state_mtgp32) * n);
@@ -395,7 +397,7 @@ rocrand_status rocrand_make_state_mtgp32(rocrand_state_mtgp32 * d_state,
     }
 
     const hipError_t error
-        = hipMemcpy(d_state, h_state, sizeof(rocrand_state_mtgp32) * n, hipMemcpyHostToDevice);
+        = hipMemcpy(state, h_state, sizeof(rocrand_state_mtgp32) * n, hipMemcpyDefault);
     free(h_state);
 
     if(error != hipSuccess)