+++ CHANGELOG.md
@@ -8,6 +8,7 @@ Documentation for rocRAND is available at
 ### Additions
 
 * Added `rocrand_create_generator_host_blocking` which dispatches without stream semantics.
+* Added host-side generator for `ROCRAND_RNG_PSEUDO_MTGP32`.
 
 ## (Unreleased) rocRAND-3.1.0 for ROCm 6.2.0